laws_database/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.10",
    "pypdf>=6.1.1",
    "sentence-transformers>=5.1.1",
    "tiktoken>=0.12.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
